% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{createAmiGOLink}
\alias{createAmiGOLink}
\title{Create AmiGO link}
\usage{
createAmiGOLink(go_term)
}
\arguments{
\item{go_term}{Go term of interest}
}
\value{
Return AmiGO link for go_term
}
\description{
Create AmiGO link
}
