% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{createEnsemblLink}
\alias{createEnsemblLink}
\title{Create Ensembl link}
\usage{
createEnsemblLink(val, ensembl_db)
}
\arguments{
\item{val}{Gene of interest}

\item{ensembl_db}{Which ensembl database to use}
}
\value{
Return NCBI link for gene of interest
}
\description{
Create Ensembl link
}
