% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{createNCBILink}
\alias{createNCBILink}
\title{Create NCBI link}
\usage{
createNCBILink(val)
}
\arguments{
\item{val}{Gene of interest}
}
\value{
Return NCBI link for gene of interest
}
\description{
Create NCBI link
}
