% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{class}
\name{expanded_OG}
\alias{expanded_OG}
\title{expanded_OG Class}
\description{
A class for expanded OG's
}
\section{Slots}{

\describe{
\item{\code{blast_table}}{df. gene level; blast hits of OG and all closest OGs}

\item{\code{add_OG_analysis}}{list. lists subclass add_OG_set elements}
}}

