% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{fn_dir}
\alias{fn_dir}
\title{Create a new tmp directory for plots, etc.}
\usage{
fn_dir(currwd = NULL)
}
\arguments{
\item{currwd}{Current working directory.}
}
\value{
Creates a new tmp directory that is deleted after exiting session.
}
\description{
Create a new tmp directory for plots, etc.
}
