% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{styleColorBar_divergent}
\alias{styleColorBar_divergent}
\title{Adding barplot overlay to DataTables}
\usage{
styleColorBar_divergent(data, color_pos, color_neg)
}
\arguments{
\item{data}{Column choice of underlying dataframe which shoud receive barplot overlay.}

\item{color_pos}{Color choice for positive values.}

\item{color_neg}{Color choice for negative values.}
}
\value{
Add styling based on choices.
}
\description{
Adding barplot overlay to DataTables
}
